+++ packages/accounts-password/password_server.js
@@ -331,7 +331,12 @@ Accounts.setPassword = function (userId, newPlaintextPassword) {
 
   Meteor.users.update(
     {_id: user._id},
-    { $unset: {'services.password.srp': 1}, // XXX COMPAT WITH 0.8.1.3
+    {
+      $unset: {
+        'services.password.srp': 1, // XXX COMPAT WITH 0.8.1.3
+        'services.password.reset': 1,
+        'services.resume.loginTokens': 1
+      },
       $set: {'services.password.bcrypt': hashPassword(newPlaintextPassword)} }
   );
 };